refactor(dashboard): decouple data fetching from wedding overview chart
